Avoid warning about missing space around literal
This fixes the following warning (for PRIu64 when used in printf):

"invalid suffix on literal; C++11 requires a space between literal and string
macro"

While just a warning on GCC, this is actually on error on Clang, and since
Bioconda builds its macOS packages using Clang, Bioconda builds currently
need a patch for samblaster.

See bioconda/bioconda-recipes#20931

This also turns the warning into an error to avoid the problem in the
future.